# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  get_task.Response:
    properties:
      canEdit:
        type: boolean
      codeToSolve:
        type: string
      description:
        type: string
      responseInfo:
        $ref: '#/definitions/response_info.ResponseInfo'
    type: object
...
  regenerate.Request:
    properties:
      noiseLevel:
        maximum: 10
        minimum: 0
        type: integer
      skipsNumber:
        minimum: 0
        type: integer
    type: object
  regenerate.Response:
    properties:
      responseInfo:
        $ref: '#/definitions/response_info.ResponseInfo'
      taskAlias:
        type: string
    type: object
  response_info.ResponseInfo:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for the Codular project, providing functionality to generate code
    skips, retrieve tasks, and check task status.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Codular API
  version: "1.0"
paths:
...
  /task/{alias}:
    get:
      description: Retrieves a task by its alias, returning the task code, description
        (title), and edit permissions for the authenticated user. Requires user authorization.
      parameters:
      - description: Task alias
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Example response" Example({"responseInfo":{"status":"OK"},"description":"String
            concatenation task","codeToSolve":"s1 + s2","canEdit":true})
          schema:
            $ref: '#/definitions/get_task.Response'
        "400":
          description: Task alias is empty
          schema:
            $ref: '#/definitions/get_task.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/get_task.Response'
        "404":
          description: Task not found or error retrieving task data
          schema:
            $ref: '#/definitions/get_task.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/get_task.Response'
      security:
      - Bearer: []
      summary: Get task by alias
      tags:
      - Tasks
  /task/{alias}/regenerate:
    patch:
      consumes:
      - application/json
      description: Regenerates an existing task (skips or noises) by its alias with
        optional new parameters (skips number or noise level). Updates the task code
        and description in the database. Requires user authorization and edit permissions.
        Returns the task alias for retrieving the updated task code and description.
      parameters:
      - description: Task alias
        in: path
        name: alias
        required: true
        type: string
      - description: Optional skips number or noise level
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/regenerate.Request'
      produces:
      - application/json
      responses:
        "200":
          description: Example response" Example({"responseInfo":{"status":"OK"},"taskAlias":"abc123"})
          schema:
            $ref: '#/definitions/regenerate.Response'
        "400":
          description: Invalid request, task alias is empty, or required parameters
            missing
          schema:
            $ref: '#/definitions/regenerate.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/regenerate.Response'
        "403":
          description: 'Forbidden: user does not have edit permissions'
          schema:
            $ref: '#/definitions/regenerate.Response'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/regenerate.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/regenerate.Response'
      security:
      - Bearer: []
      summary: Regenerate task by alias
      tags:
      - Tasks
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"